requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.1
numpy==1.26.4
pydantic==2.5.0
python-multipart==0.0.6
torchvision
Pillow
opencv-python
dlib
scipy
joblib
matplotlib
seaborn
scikit-learn
tensorflow